refactor(subscriberdb): Segregation of SubscriberDBCloud servicer
